qt-creator: Fix installation paths and install desktop file
Patch qt-creator in order to support the build API which gives us the
chance to pass a couple of arguments to install into /usr instead of /
and libraries to /usr/lib instead of /lib.

Also, install desktop file so that users will be able to run QtCreator
from the desktop environment.