qtbase: Change installation paths
Since we are using qtchooser we install Qt 5 in dedicated paths.
This will help us having two incompatible versions at the same time,
for example both Qt 5 and Qt 6 during the transition.

I know this won't happen soon but it's better to get this right now,
otherwise when we'll get more users this will be painful.